bti: incorrect version fix
bti_version.h is not getting updated properly which results in incorrect
version name.

This bug happens only if the previous build is not cleaned and the files are
updated to newer version.

This following patch fixes the issue [not sure is the right way to do it] by
making the bti_version.h file depends on Makefile since the version details
are added to Makefile

I have tested this patch with different versions of bti and are working  fine

Signed-off-by: SanjayKumar J <vjsanjay@gmail.com>