add bti --shrink-urls
This patch adds a bti-shrink-urls script which uses http://2tu.us/ to
convert any URLs to something more sensible for a tweet.

bti, with the --shrink-urls option, now uses the above to convert any
URLs it finds.

'shrink-urls' can also be set to 'true' or 'yes' in the config file.

Signed-off-by: Bart Trojanowski <bart@jukie.net>