Support for custom bti installations
Now one can pass the URL of a custom laconica installation as value of
host (the both previous "twitter" and "identica" are still valid
though). The passed URL should be the base path of where API endpoints
are located. e.g., identica' baseurl is:
  http://identi.ca/api/statuses
And twit.tv's army (which is a laconica instance) is:
  http://army.twit.tv/api/statuses