Remove link-time dependancy on readline
This patch divorces bti from depending on readline.  It now
will try to work with any readline-like library on your system,
and if none is found, fall back to an internal function.

This will solve the problems that distros are having with the
readline-6 license change to GPLv3 and bti staying at GPLv2.

Logic for this is taken from e2fsprogs, which implemented this
a long time ago for the same type of reason.

Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>