Fixing Identi.ca's OAuth request_token URI
Revision "a" of OAuth 1.0 mandates the presence of "oauth_callback"
parameter.

Usually libraries appends an "oob" (Out of Band) oauth_callback if you
don't provide a valid callback url, and don't specify that you don't
want "OAuth 1.0a".

But apparently our library of choice for OAuth (liboauth) does not do
this, though the code is in library but commented out.

src/oauth.c:675-680 (as in HEAD, and 0.9.0 tarball):

675 #if 0 // oauth_version 1.0 Rev A
676   if (!oauth_param_exists(argv,argc,"oauth_callback")) {
677     snprintf(oarg, 1024, "oauth_callback=oob");
678     oauth_add_param_to_array(argcp, argvp, oarg);
679   }
680 #endif

So for now, we have to add this to our request_token URI.